refactor(clube-event): simplifica nomenclatura de ClubeEventType e separa lógica da controller
- Renomeia valores do enum ClubeEventType de CRIACAO_CLUBE, ATUALIZACAO_CLUBE, EXCLUSAO_CLUBE para CRIACAO, ATUALIZACAO, EXCLUSAO para evitar redundância
- Altera pasta util que tinha enum ClubeEventType para enums, organizando melhor a estrutura de enums do projeto
- Cria CriarClubeEventService para orquestrar montagem do ClubeEvent e chamada do CriarClubeProducer, removendo lógica de montagem da controller
- Atualiza CriarClubeController para delegar processamento para CriarClubeEventService, eliminando criação direta de ClubeEvent no controller
- Garante desacoplamento e separação de responsabilidades entre controller, service e producer
