+++ partidas-de-futebol-api/src/main/java/br/com/meli/teamcubation_partidas_de_futebol/clube/rabbitmq/enums/ClubeEventStatus.java
@@ -1,0 +1,8 @@
+package br.com.meli.teamcubation_partidas_de_futebol.clube.rabbitmq.enums;
+
+public enum ClubeEventStatus {
+    CREATED,
+    UPDATED,
+    DELETED,
+
+}
