+++ partidas-de-futebol-api/src/main/java/br/com/meli/teamcubation_partidas_de_futebol/clube/rabbitmq/enums/ClubeEventType.java
@@ -1,0 +1,7 @@
+package br.com.meli.teamcubation_partidas_de_futebol.clube.rabbitmq.enums;
+
+public enum ClubeEventType {
+    CRIACAO,
+    ATUALIZACAO,
+    EXCLUSAO
+}
