--- partidas-de-futebol-api/src/main/java/br/com/meli/teamcubation_partidas_de_futebol/clube/rabbitmq/util/ClubeEventStatus.java
@@ -1,8 +1,0 @@
-package br.com.meli.teamcubation_partidas_de_futebol.clube.rabbitmq.util;
-
-public enum ClubeEventStatus {
-    CREATED,
-    UPDATED,
-    DELETED,
-
-}
