--- partidas-de-futebol-api/src/main/java/br/com/meli/teamcubation_partidas_de_futebol/clube/rabbitmq/util/ClubeEventType.java
@@ -1,7 +1,0 @@
-package br.com.meli.teamcubation_partidas_de_futebol.clube.rabbitmq.util;
-
-public enum ClubeEventType {
-    CRIACAO_CLUBE,
-    ATUALIZACAO_CLUBE,
-    EXCLUSAO_CLUBE
-}
