+++ partidas-de-futebol-mensageria/src/main/java/br/com/meli/partidas_de_futebol_mensageria/clube/enums/ClubeEventStatus.java
@@ -1,4 +1,4 @@
-package br.com.meli.partidas_de_futebol_mensageria.clube.util;
+package br.com.meli.partidas_de_futebol_mensageria.clube.enums;
 
 public enum ClubeEventStatus {
     CREATED,
