+++ partidas-de-futebol-mensageria/src/main/java/br/com/meli/partidas_de_futebol_mensageria/clube/enums/ClubeEventType.java
@@ -1,0 +1,7 @@
+package br.com.meli.partidas_de_futebol_mensageria.clube.enums;
+
+public enum ClubeEventType {
+    CRIACAO,
+    ATUALIZACAO,
+    EXCLUSAO
+}
