--- partidas-de-futebol-mensageria/src/main/java/br/com/meli/partidas_de_futebol_mensageria/clube/util/ClubeEventType.java
@@ -1,7 +1,0 @@
-package br.com.meli.partidas_de_futebol_mensageria.clube.util;
-
-public enum ClubeEventType {
-    CRIACAO_CLUBE,
-    ATUALIZACAO_CLUBE,
-    EXCLUSAO_CLUBE
-}
